new index.ts file for course json files
