+++ .prettierignore
@@ -1,1 +1,1 @@
-/src/courses
+/src/courses/*.json
